Manage query string

+++ proxy.rb
@@ -7,10 +7,16 @@ class Proxy < Sinatra::Base
   get '/*/*' do
     host = params[:splat][0]
     path = "/" + params[:splat][1]
+    path += "?#{request.query_string}" if request.query_string && !request.query_string.empty?
     puts "HTTP Call #{host} #{path}"
+    target = File.join(File.dirname(__FILE__), "storage", "200", host, path)
+    if File.exists? target
+      content_type 'application/octet-stream'
+      response.write File.read(target)
+      return 200
+    end
     result = Net::HTTP.get_response(host, path)
     if result.code == "200"
-      target = File.join(File.dirname(__FILE__), "storage", "200", host, path)
       FileUtils.mkdir_p File.dirname(target)
       File.open(target, "wb") {|io| io.write result.body}
       content_type 'application/octet-stream'
